\name{Clustering Algorithms}
\alias{clustering_algorithms}
\alias{cluster_edge_betweenness}
\alias{cluster_fast_greedy}
\alias{cluster_fluid_communities}
\alias{cluster_infomap}
\alias{cluster_label_prop}
\alias{cluster_leading_eigen}
\alias{cluster_leiden}
\alias{cluster_louvain}
\alias{cluster_optimal}
\alias{cluster_spinglass}
\alias{cluster_walktrap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Clustering Algorithms}
\description{
Clustering algorithm functions available for use within the \code{NAIR} package. Used as arguments to functions such as \code{\link{buildRepSeqNetwork}} and \code{\link{addClusterMembership}}.

These functions have been imported from the \code{\link[igraph]{igraph}} package. See the original documentation linked below for more details.
}
\seealso{
\code{\link[igraph]{cluster_edge_betweenness}}
\code{\link[igraph]{cluster_fast_greedy}}
\code{\link[igraph]{cluster_fluid_communities}}
\code{\link[igraph]{cluster_infomap}}
\code{\link[igraph]{cluster_label_prop}}
\code{\link[igraph]{cluster_leading_eigen}}
\code{\link[igraph]{cluster_leiden}}
\code{\link[igraph]{cluster_louvain}}
\code{\link[igraph]{cluster_optimal}}
\code{\link[igraph]{cluster_spinglass}}
\code{\link[igraph]{cluster_walktrap}}
}
\author{
%%  ~~who you are~~
    Brian Neal (\email{Brian.Neal@ucsf.edu})
}