\name{generateNetworkGraphPlots}
\alias{generateNetworkGraphPlots}
\title{
  Generate Plots of a Network Graph
}
\description{
  Given the \code{\link[igraph]{igraph}} of a network and a data frame containing the corresponding node-level meta data, generates one or more \code{\link[ggraph]{ggraph}} plots of the network graph according to the user specifications.
}
\usage{
generateNetworkGraphPlots(
  igraph, data, print_plots = TRUE,
  plot_title = NULL, plot_subtitle = NULL,
  color_nodes_by = NULL, color_scheme = "default",
  color_legend = "auto", color_title = "auto",
  edge_width = 0.1, size_nodes_by = 0.5,
  node_size_limits = NULL, size_title = "auto")
}

\arguments{

\item{igraph}{
  An \code{\link[igraph]{igraph}} object containing the network graph to be plotted.
}

\item{data}{
  A data frame containing the node-level meta-data for the network, with each row corresponding to a node.
}

\item{print_plots}{
  A logical scalar; should plots be printed in the \code{R} plotting window?
}

\item{plot_title}{
  Passed to \code{\link{plotNetworkGraph}} when generating each plot.
}

\item{plot_subtitle}{
  Passed to \code{\link{plotNetworkGraph}} when generating each plot.
}

\item{color_nodes_by}{
  A vector specifying one or more variables of \code{data} used to encode the color of the nodes. One plot is generated for each vector entry, with each plot coloring the nodes according to the variable in the corresponding vector entry. The vector must be either a character vector (in which case each entry is a column name of \code{data}) or a numeric vector (in which case each entry is a column index of \code{data}). If this argument is \code{NULL}, one plot will be generated with the nodes uncolored.
}

\item{color_scheme}{
  A character string specifying the color scale to use for all plots, or a character vector whose length matches that of \code{color_nodes_by}, specifying the color scale used for each plot. Each entry is passed to \code{\link{plotNetworkGraph}} when generating the plot for the corresponding entry of \code{color_nodes_by}.
}

\item{color_legend}{
  Passed to \code{\link{plotNetworkGraph}} when generating each plot.
}

\item{color_title}{
  A character string specifying the title of the color legend in all plots, or a character vector whose length matches that of \code{color_nodes_by}, specifying the title of the color legend in each plot. Each entry is passed to \code{\link{plotNetworkGraph}} when generating the plot for the corresponding entry of \code{color_nodes_by}.
}

\item{edge_width}{
  Passed to \code{\link{plotNetworkGraph}} when generating each plot.
}

\item{size_nodes_by}{
  A numeric scalar specifying the size of the nodes in all plots, or the column name of a variable of \code{data} used to encode the size of the nodes in all plots. Alternatively, an argument value of \code{NULL} uses the default \code{\link[ggraph]{ggraph}} size for all nodes. Passed to the size aesthetic mapping of \code{\link[ggraph]{geom_node_point}}.
}

\item{node_size_limits}{
  Passed to \code{\link{plotNetworkGraph}} when generating each plot.
}

\item{size_title}{
  Passed to \code{\link{plotNetworkGraph}} when generating each plot.
}

}

\details{
  Intended for use with the output of \code{\link{buildRepSeqNetwork}} or \code{\link{generateNetworkObjects}}. The output of these functions contain the inputs for the \code{igraph} and \code{data} arguments of \code{generateNetworkGraphPlots}.

  Each plot is generated by dispatching a call to \code{\link{plotNetworkGraph}}.

  The plots can be saved to a single pdf using \code{\link{saveNetworkPlots}}.
}

\value{
  A list containing one element for each plot generated. Each element is a \code{ggraph} object whose name is the column name of the variable used to color the nodes. If a plot has uncolored nodes, its name is \code{uniform_color}.
}

\references{
Hai Yang, Jason Cham, Brian Neal, Zenghua Fan, Tao He and Li Zhang. (2023). NAIR: Network Analysis of Immune Repertoire. \emph{Frontiers in Immunology}, vol. 14. \url{https://doi.org/10.3389/fimmu.2023.1181825}

\href{https://mlizhangx.github.io/Network-Analysis-for-Repertoire-Sequencing-/index.html}{Webpage for the NAIR package}

\href{https://mlizhangx.github.io/Network-Analysis-for-Repertoire-Sequencing-/articles/network_visualization.html}{Network Visualization vignette}
}

\author{
    Brian Neal (\email{Brian.Neal@ucsf.edu})
}

\seealso{
  \code{\link{saveNetworkPlots}}

  \code{\link{generateNetworkObjects}}

  \code{\link{plotNetworkGraph}}
}

\examples{
\dontrun{
# Simulate some toy data for demonstration
toy_data <- simulateToyData()

# Generate network for data
net <- generateNetworkObjects(toy_data, "CloneSeq")

# Plot network graph
net_plot <- generateNetworkGraphPlots(
  net$igraph, net$node_data, color_nodes_by = c("SampleID", "CloneCount"))
}
}
