\docType{package}
\name{NAIR-package}
\alias{NAIR}
\alias{NAIR-package}
\title{NAIR: Network Analysis of Immune Repertoire}
\description{
To learn about the NAIR package and get started, visit the
\href{https://mlizhangx.github.io/Network-Analysis-for-Repertoire-Sequencing-/}{package website},
or browse the package vignettes offline:

\code{browseVignettes(package = "NAIR")}

The following vignette is a good place to start:

\code{vignette("NAIR", package = "NAIR")}
}
\seealso{
\itemize{
  \item \href{https://mlizhangx.github.io/Network-Analysis-for-Repertoire-Sequencing-/}{Package website}
  \item \href{https://github.com/mlizhangx/Network-Analysis-for-Repertoire-Sequencing-}{Github page}
  \item \href{https://github.com/mlizhangx/Network-Analysis-for-Repertoire-Sequencing-/issues}{Report bugs and issues here}
}
}

\author{
\itemize{
  \item Brian Neal (\email{Brian.Neal@ucsf.edu}), Maintainer
  \item Hai Yang (\email{Hai.Yang@ucsf.edu})
  \item Phi-Long Le (\email{PhiLong.Le@ucsf.edu})
  \item Li Zhang (\email{Li.Zhang@ucsf.edu})
}

}